Don't pull in all of nautilus for the -extensions subpackage
The -extensions subpackage that ships libnautilus-extension.so.1
currently has a hard dep on the main nautilus subpackage. This is
problematic because it causes anything that links with
libnautilus-extension.so.1 to automatically pull in all of nautilus.

With the current appdata and gnome-software setup, we don't want to have
one gui app ever depending on another gui app.

As a fix, just drop the nautilus dep from nautilus-extensions and allow
the libraries to be installed standalone.